fixed recursive dir iterator

+++ lib/cache/sfFileCache.class.php
@@ -247,9 +247,12 @@ class sfFileCache extends sfCache {
     }
 
     $result = true;
-    foreach(new RecursiveIteratorIterator(new RecursiveDirectoryIterator($this->getOption('cache_dir'),            
-            FilesystemIterator::SKIP_DOTS)) as $file)
+    foreach(new RecursiveIteratorIterator(new RecursiveDirectoryIterator($this->getOption('cache_dir'))) as $file)
     {
+      if($file->isDot())
+      {
+        continue;
+      }
       if(self::MODE_ALL == $mode || !$this->isValid($file))
       {
         $result = @unlink($file) && $result;
